HOP-3249: re enable and fix some unit tests
