+++ plugins/engines/beam/src/test/java/org/apache/hop/beam/engines/direct/BeamDirectPipelineEngineTest.java
@@ -24,7 +24,6 @@ import org.apache.hop.pipeline.PipelineMeta;
 import org.apache.hop.pipeline.config.IPipelineEngineRunConfiguration;
 import org.apache.hop.pipeline.config.PipelineRunConfiguration;
 import org.apache.hop.pipeline.engine.IPipelineEngine;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.Arrays;
@@ -34,7 +33,6 @@ import static org.junit.Assert.assertEquals;
 public class BeamDirectPipelineEngineTest extends BeamBasePipelineEngineTest {
 
   @Test
-  @Ignore
   public void testDirectPipelineEngine() throws Exception {
 
     IPipelineEngineRunConfiguration configuration = new BeamDirectPipelineRunConfiguration();
